backend/requirements.txt
Flask==0.12
flask-restful==0.3.5
google_api_python_client==1.5.3